Create Root.Conjunction to replace RootResolver
